floats!

+++ tsp_local/main.cpp
@@ -7,9 +7,9 @@
 using namespace std;
 
 
-int distance(pair<int,int> &p1, pair<int,int> &p2)
+float distance(pair<float,float> &p1, pair<float,float> &p2)
 {
-    int dist;
+    float dist;
     
     dist = sqrt( pow((p1.first - p2.first),2) + pow((p1.second - p2.second),2) );
     return dist;
@@ -22,20 +22,20 @@ int main() {
     cin>> n;
     //we need a shortest possible route that connects every vertex
     
-    int x,y;
-    multimap<int,int> points;
+    float x,y;
+    multimap<float,float> points;
     
     for(int i = 0; i<n;i++)
     {
         cin>> x;
         cin>> y;
-        points.insert(pair<int, int> (x,y));
+        points.insert(pair<float, float> (x,y));
         
     } //automatically ordered by x-coordinate
     
     
     
-    multimap<int, int>::iterator itr;
+    multimap<float, float>::iterator itr;
     for (itr = points.begin(); itr != points.end(); ++itr) {
             cout << '\t' << itr->first << '\t' << itr->second
                  << '\n';
